feat(ios): still enable metal backend on simulator & build metallib for simulator (#120)

+++ scripts/build-ios.sh
@@ -43,7 +43,11 @@ function build_framework() {
   # Copy headers and metallib
   cp_headers $4
   # TODO: May need to re-build metallib for tvOS
-  cp ../cpp/ggml-llama.metallib ../ios/rnllama.xcframework/$4/rnllama.framework/ggml-llama.metallib
+  if [ "$4" == "ios-arm64_x86_64-simulator" ] || [ "$4" == "tvos-arm64_x86_64-simulator" ]; then
+    cp ../cpp/ggml-llama-sim.metallib ../ios/rnllama.xcframework/$4/rnllama.framework/ggml-llama-sim.metallib
+  else
+    cp ../cpp/ggml-llama.metallib ../ios/rnllama.xcframework/$4/rnllama.framework/ggml-llama.metallib
+  fi
 
   rm -rf ./*
   cd ..
